# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "RAG System API",
    "description": "Ollama와 LangGraph 기반의 고도화된 RAG 시스템 API",
    "version": "2.0.0"
  },
  "paths": {
    "/api/v1/health": {
      "get": {
        "summary": "Health Check",
        "description": "서버 상태 확인",
        "operationId": "health_check_api_v1_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/api/v1/upload": {
      "post": {
        "summary": "Upload Document",
        "description": "PDF 문서를 업로드하고 해당 세션에 인덱싱합니다.",
        "operationId": "upload_document_api_v1_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_document_api_v1_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/query": {
      "post": {
        "summary": "Query Rag",
        "description": "격리된 세션 컨텍스트에서 질의를 수행합니다.",
        "operationId": "query_rag_api_v1_query_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/stream_query": {
      "post": {
        "summary": "Stream Query Rag",
        "description": "실시간 스트리밍(SSE) 응답 시 세션 격리를 보장합니다.",
        "operationId": "stream_query_rag_api_v1_stream_query_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Body_upload_document_api_v1_upload_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          },
          "session_id": {
            "type": "string",
            "title": "Session Id",
            "default": "default"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_document_api_v1_upload_post"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "QueryRequest": {
        "properties": {
          "query": {
            "type": "string",
            "title": "Query",
            "example": "DeepSeek-R1의 성능은 어때?"
          },
          "session_id": {
            "type": "string",
            "title": "Session Id",
            "default": "default",
            "example": "user-123"
          },
          "use_cache": {
            "type": "boolean",
            "title": "Use Cache",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "query"
        ],
        "title": "QueryRequest",
        "description": "질의 요청 스키마"
      },
      "QueryResponse": {
        "properties": {
          "answer": {
            "type": "string",
            "title": "Answer"
          },
          "sources": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Sources",
            "default": []
          },
          "execution_time_ms": {
            "type": "number",
            "title": "Execution Time Ms"
          }
        },
        "type": "object",
        "required": [
          "answer",
          "execution_time_ms"
        ],
        "title": "QueryResponse",
        "description": "질의 응답 스키마"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}